rag-financial-docs/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-financial-docs"
version = "0.1.0"
description = "RAG system for financial and legal documents — built from scratch and with LangChain"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # LLM & Embeddings
    "openai>=1.30.0",
    "tiktoken>=0.7.0",

    # LangChain ecosystem
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-qdrant>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-huggingface>=0.0.3",
    "langgraph>=0.1.0",

    # Vector DB
    "qdrant-client>=1.9.0",

    # PDF parsing
    "pymupdf>=1.24.0",        # fitz — fast PDF text extraction
    "pdfplumber>=0.11.0",     # table extraction from PDFs

    # DOCX parsing
    "python-docx>=1.1.0",

    # Sparse search
    "rank-bm25>=0.2.2",

    # Re-ranking
    "sentence-transformers>=3.0.0",
    "torch>=2.3.0",

    # Evaluation
    "ragas>=0.1.9",
    "datasets>=2.19.0",

    # Synthetic data generation
    "reportlab>=4.2.0",       # Generate PDFs programmatically
    "faker>=25.0.0",          # Fake names, companies, addresses

    # Config & env
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",

    # UI
    "streamlit>=1.35.0",

    # Utilities
    "tenacity>=8.3.0",        # Retry logic for API calls
    "tqdm>=4.66.0",
    "rich>=13.7.0",           # Pretty terminal output
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
